Adding check for Enum in IsSimpleType
Without this check `ExpressionReader.ReadParameters` assumes it's a complex type, and tries to read the public properties of the Enum type. Since there aren't any public properties on an Enum, the call to `values.AddRange` in line 176 of the previously mentioned method won't add anything and the Enum parameter value will be lost. This change should fix that.